chore(wave2): clean up lobby and tween warnings
